@@ -4,33 +4,33 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build --prod",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "start": "ng serve dev",
+    "prod:start": "ng serve dev --prod",
+    "build": "ng build dev --prod",
+    "test": "ng test dev",
+    "lint": "ng lint dev"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.0",
-    "@angular/common": "5.2.0",
-    "@angular/compiler": "5.2.0",
-    "@angular/core": "5.2.0",
-    "@angular/forms": "5.2.0",
-    "@angular/http": "5.2.0",
-    "@angular/platform-browser": "5.2.0",
-    "@angular/platform-browser-dynamic": "5.2.0",
-    "@angular/router": "5.2.0",
+    "@angular/animations": "6.0.1",
+    "@angular/common": "6.0.1",
+    "@angular/compiler": "6.0.1",
+    "@angular/core": "6.0.1",
+    "@angular/forms": "6.0.1",
+    "@angular/http": "6.0.1",
+    "@angular/platform-browser": "6.0.1",
+    "@angular/platform-browser-dynamic": "6.0.1",
+    "@angular/router": "6.0.1",
     "bpmn-js": "1.2.1",
     "bpmn-js-task-priorities": "0.2.0",
     "core-js": "2.4.1",
-    "rxjs": "5.5.6",
-    "zone.js": "0.8.19"
+    "rxjs": "6.1.0",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "^6.0.1",
+    "@angular/compiler-cli": "6.0.1",
+    "@angular/language-service": "6.0.1",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -45,6 +45,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.1"
   }
 }
