@@ -1,16 +1,10 @@
+import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BpmnViewerComponent } from './bpmn-viewer/bpmn-viewer.component';
-import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
-    imports: [
-        HttpClientModule
-    ],
-    declarations: [
-        BpmnViewerComponent
-    ],
-    exports: [
-        BpmnViewerComponent
-    ]
+  imports: [HttpClientModule],
+  declarations: [BpmnViewerComponent],
+  exports: [BpmnViewerComponent]
 })
-export class BpmnModule { }
+export class BpmnModule {}
